Move controller pdf method to a concern
