+++ app/controllers/numismatics/coins_controller.rb
@@ -13,6 +13,8 @@ module Numismatics
     before_action :load_numismatic_collections, only: [:new, :edit, :update]
     before_action :selected_issue, only: [:new, :edit, :destroy, :update]
 
+    include Pdfable
+
     def facet_fields
       [
         :numismatic_collection_ssim
@@ -73,19 +75,6 @@ module Numismatics
         elsif params[:parent_id]
           find_resource(params[:parent_id])
         end
-    end
-
-    def pdf
-      change_set = ChangeSet.for(find_resource(params[:id]))
-      authorize! :pdf, change_set.resource
-      resource_id = change_set.resource.id
-      GeneratePdfJob.perform_now(resource_id: resource_id)
-
-      pdf_file = query_service.find_by(id: resource_id).decorate.pdf_file
-
-      redirect_path_args = { resource_id: change_set.id, id: pdf_file.id }
-      redirect_path_args[:auth_token] = auth_token_param if auth_token_param
-      redirect_to download_path(redirect_path_args)
     end
 
     def selected_issue
